+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/test_00_prepare.py
@@ -1,4 +1,4 @@
-from unittest.mock import ANY, Mock, patch
+from unittest.mock import Mock, patch
 
 import projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate._00_prepare as job
 from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
@@ -13,8 +13,7 @@ class TestPrepare:
     PREPARED_DATA_DESTINATION = "some/destination"
 
     @patch(f"{PATCH_PATH}.utils.sink_to_parquet")
-    @patch(f"{PATCH_PATH}.apply_categorical_labels")
-    @patch(f"{PATCH_PATH}.pUtils.convert_job_role_strings_to_number_only")
+    @patch(f"{PATCH_PATH}.pUtils.relabel_job_role_columns")
     @patch(f"{PATCH_PATH}.pUtils.pivot_job_role_cols_to_rows")
     @patch(f"{PATCH_PATH}.pUtils.reduce_to_published_roles")
     @patch(f"{PATCH_PATH}.earliest_file_per_month_filter_expr")
@@ -29,8 +28,7 @@ class TestPrepare:
         earliest_file_per_month_filter_expr_mock: Mock,
         reduce_to_published_roles_mock: Mock,
         pivot_job_role_cols_to_rows_mock: Mock,
-        convert_job_role_strings_to_number_only_mock: Mock,
-        apply_categorical_labels_mock: Mock,
+        relabel_job_role_columns_mock: Mock,
         sink_to_parquet_mock: Mock,
     ):
         job.main(
@@ -63,16 +61,16 @@ class TestPrepare:
             earliest_file_per_month_filter_expr_mock.return_value
         )
 
-        # TODO: Uncomment these assertions when the placeholder functions are implemented
         reduce_to_published_roles_mock.assert_called_once()
         merged_jr_cols_lf = reduce_to_published_roles_mock.return_value
         merged_jr_cols_lf.drop.assert_called_once()
         dropped_cols_lf = merged_jr_cols_lf.drop.return_value
         # pivot_job_role_cols_to_rows_mock.assert_called_once()
-        # convert_job_role_strings_to_number_only_mock.assert_called_once()
-        # apply_categorical_labels_mock.assert_called_once()
+
+        relabel_job_role_columns_mock.assert_called_once_with(dropped_cols_lf)
+        relabelled_lf = relabel_job_role_columns_mock.return_value
 
         sink_to_parquet_mock.assert_called_once_with(
-            lazy_df=dropped_cols_lf,
+            lazy_df=relabelled_lf,
             output_path=self.PREPARED_DATA_DESTINATION,
         )